Function to execute a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int j;
+	size_t j;
 
 	if (!(array && action))
 		return;
